added test for predict endpoint
